+++ src/test_entities.rs
@@ -4,11 +4,11 @@ use crate::entity_event_engine::TimedEntity;
 #[derive(Debug, Clone, Default)]
 pub struct TestEntity {
     pub name: String,
-    pub speed: f32
+    pub speed: f64
 }
 
 impl TimedEntity for TestEntity {
-    fn get_speed(&self) -> f32 {
+    fn get_speed(&self) -> f64 {
         self.speed
     }
     fn update(&self) {
@@ -19,12 +19,12 @@ impl TimedEntity for TestEntity {
 #[derive(Debug, Clone, Default)]
 pub struct TestEntity2 {
     pub name: String,
-    pub speed: f32,
-    pub whatever: f32
+    pub speed: f64,
+    pub whatever: f64
 }
 
 impl TimedEntity for TestEntity2 {
-    fn get_speed(&self) -> f32 {
+    fn get_speed(&self) -> f64 {
         self.speed
     }
     fn update(&self) {
